Tests on Jetson, send back to PC for plotting

+++ gen_tests/disparity.sh
@@ -7,14 +7,17 @@ INTERF=" "
 sysctl -w kernel.sched_rt_runtime_us=-1
 
 # Make Memguard (if necessary)
-echo "Building Memguard"
-(cd ../patched_memguard && make)
+# echo "Building Memguard"
+# (cd ../patched_memguard && make)
 
 
 echo "$TEST in Isolation"
 INTERF="Isolated"
 
 sudo insmod ../patched_memguard/memguard.ko g_hw_counter_id=0x17
+sleep 2
+
+
 sudo echo 0 > /sys/kernel/debug/tracing/trace
 
 taskset -c 0 ../benchmarks/sd-vbs/benchmarks/disparity/data/fullhd/disparity ../benchmarks/sd-vbs/benchmarks/disparity/data/fullhd/. | grep "Cycles elapsed" > cycles_${TEST}${INTERF}.txt & PID_TO_WAIT=$!
@@ -27,8 +30,8 @@ python3 splitftrace.py trace_${TEST}${INTERF}.txt
 
 
 
-
-sleep 5
+echo "done"
+sleep 10
 
 
 
@@ -38,6 +41,7 @@ INTERF="hesocMemcpy"
 echo "$TEST against $INTERF"
 
 sudo insmod ../patched_memguard/memguard.ko g_hw_counter_id=0x17
+sleep 2
 
 sudo taskset -c 0 ../benchmarks/hesoc-mark/cuda/cudainterf -d 102400 -i 20000 -m memcpy & PID_TO_KILL0=$!
 
@@ -48,8 +52,22 @@ taskset -c 0 ../benchmarks/sd-vbs/benchmarks/disparity/data/fullhd/disparity ../
 wait $PID_TO_WAIT
 sudo rmmod memguard
 
-sudo killall -9 cudainterf
-sudo kill -9 $PID_TO_KILL0
+if ps -p $PID_TO_KILL0 > /dev/null
+then
+	echo "still running"
+	sudo killall -9 cudainterf
+fi
+
+if ps -p $PID_TO_KILL0 > /dev/null
+then
+	echo "STILL running"
+	sudo kill -9 $PID_TO_KILL0
+fi
+
+if ps -p $PID_TO_KILL0 > /dev/null
+then
+	echo "STILL RUNNING"
+fi
 
 sudo cat /sys/kernel/debug/tracing/trace > trace_${TEST}_vs_${INTERF}.txt 
 
@@ -59,7 +77,8 @@ python3 splitftrace.py trace_${TEST}_vs_${INTERF}.txt
 
 
 
-sleep 5
+echo "done"
+sleep 10
 
 
 
@@ -68,6 +87,7 @@ INTERF="hesocMemset"
 echo "$TEST against $INTERF"
 
 sudo insmod ../patched_memguard/memguard.ko g_hw_counter_id=0x17
+sleep 2
 
 sudo taskset -c 0 ../benchmarks/hesoc-mark/cuda/cudainterf -d 102400 -i 20000 -m memset & PID_TO_KILL0=$!
 
@@ -78,8 +98,22 @@ taskset -c 0 ../benchmarks/sd-vbs/benchmarks/disparity/data/fullhd/disparity ../
 wait $PID_TO_WAIT
 sudo rmmod memguard
 
-sudo killall -9 cudainterf
-sudo kill -9 $PID_TO_KILL0
+if ps -p $PID_TO_KILL0 > /dev/null
+then
+	echo "still running"
+	sudo killall -9 cudainterf
+fi
+
+if ps -p $PID_TO_KILL0 > /dev/null
+then
+	echo "STILL running"
+	sudo kill -9 $PID_TO_KILL0
+fi
+
+if ps -p $PID_TO_KILL0 > /dev/null
+then
+	echo "STILL RUNNING"
+fi
 
 sudo cat /sys/kernel/debug/tracing/trace > trace_${TEST}_vs_${INTERF}.txt 
 
@@ -88,7 +122,8 @@ python3 splitftrace.py trace_${TEST}_vs_${INTERF}.txt
 
 
 
-sleep 5
+echo "done"
+sleep 10
 
 
 
@@ -98,6 +133,7 @@ INTERF="bandwidthRead"
 echo "$TEST against $INTERF"
 
 sudo insmod ../patched_memguard/memguard.ko g_hw_counter_id=0x17
+sleep 2
 
 for c in 1 2 3; do bandwidth -c $c -t 1000 & done
 
@@ -118,7 +154,8 @@ python3 splitftrace.py trace_${TEST}_vs_${INTERF}.txt
 
 
 
-sleep 5
+echo "done"
+sleep 10
 
 
 
@@ -128,6 +165,7 @@ INTERF="bandwidthWrite"
 echo "$TEST against $INTERF"
 
 sudo insmod ../patched_memguard/memguard.ko g_hw_counter_id=0x17
+sleep 2
 
 for c in 1 2 3; do bandwidth -a write -c $c -t 1000 & done
 
@@ -149,7 +187,8 @@ python3 splitftrace.py trace_${TEST}_vs_${INTERF}.txt
 
 
 
-sleep 5
+echo "done"
+sleep 10
 
 
 
@@ -160,6 +199,7 @@ INTERF="bandwidthRead_hesocMemcpy"
 echo "$TEST against $INTERF"
 
 sudo insmod ../patched_memguard/memguard.ko g_hw_counter_id=0x17
+sleep 2
 
 for c in 1 2 3; do bandwidth -c $c -t 1000 & done
 sudo taskset -c 0 ../benchmarks/hesoc-mark/cuda/cudainterf -d 102400 -i 20000 -m memcpy & PID_TO_KILL0=$!
@@ -183,7 +223,8 @@ python3 splitftrace.py trace_${TEST}_vs_${INTERF}.txt
 
 
 
-sleep 5
+echo "done"
+sleep 10
 
 
 
@@ -195,6 +236,7 @@ INTERF="bandwidthWrite_hesocMemset"
 echo "$TEST against $INTERF"
 
 sudo insmod ../patched_memguard/memguard.ko g_hw_counter_id=0x17
+sleep 2
 
 for c in 1 2 3; do bandwidth -a write -c $c -t 1000 & done 
 sudo taskset -c 0 ../benchmarks/hesoc-mark/cuda/cudainterf -d 102400 -i 20000 -m memset & PID_TO_KILL0=$!
@@ -217,13 +259,14 @@ python3 splitftrace.py trace_${TEST}_vs_${INTERF}.txt
 
 
 
-
-
-
-
-
-
-mv *.txt tests/
 echo "done"
 
-(cd ../patched_memguard && make clean)
+
+
+
+
+
+mv *.txt tests_uSec/
+echo "done"
+
+# (cd ../patched_memguard && make clean)
